move ErrorResponse json_models
